+++ api/code/firewall.go
@@ -153,6 +153,7 @@ var WireguardSocketPath = TEST_PREFIX + "/state/plugins/wireguard/wireguard_plug
 var BASE_READY = TEST_PREFIX + "/state/base/ready"
 
 var DEVICE_POLICY_PERMIT_PRIVATE_UPSTREAM_ACCESS = "lan_upstream"
+var DEVICE_POLICY_NOAPI = "noapi"
 
 const firstOutboundRouteTable = 11
 
@@ -380,10 +381,12 @@ func getDefaultGatewayLocked(dev string) (string, error) {
 	return "", fmt.Errorf("gateway not found")
 }
 
-func setDefaultUplinkGateway(iface string, index int) {
+// false -> did not set Gateway
+// true -> have a gateway or did not set one
+func setDefaultUplinkGateway(iface string, index int) bool {
 	// do not mess with route for mesh for now
 	if isLeafRouter() {
-		return
+		return false
 	}
 
 	gateway, err := getDefaultGatewayLocked(iface)
@@ -391,9 +394,9 @@ func setDefaultUplinkGateway(iface string, index int) {
 		//no gateway found, continue on
 		if err != nil {
 			//only log when err is not nil
-			log.Println("failed to set default gw for "+iface+": not found", err)
+			log.Println("failed to get default gw for "+iface+": not found", err)
 		}
-		return
+		return false
 	}
 
 	table := fmt.Sprintf("%d", firstOutboundRouteTable+index)
@@ -401,17 +404,20 @@ func setDefaultUplinkGateway(iface string, index int) {
 	current_table_route := getRouteGatewayForTable(table)
 	if current_table_route == gateway {
 		// route already set, make no updates
-		return
+		return true
 	}
 
+	ret := true
 	cmd := exec.Command("ip", "route", "replace", "default", "via", gateway, "dev", iface, "table", table)
 	_, err = cmd.Output()
 	if err != nil {
 		log.Print("Error with route setup", cmd, err)
+		ret = false
 	}
 
 	cmd = exec.Command("ip", "route", "flush", "cache")
 	_, _ = cmd.Output()
+	return ret
 }
 
 func updateOutboundRoutes() {
@@ -442,6 +448,13 @@ func collectOutbound() []string {
 	outbound := []string{}
 	for _, iface := range interfaces {
 		if iface.Type == "Uplink" && iface.Subtype != "pppup" && iface.Enabled {
+
+			gw, _ := getDefaultGatewayLocked(iface.Name)
+			if gw == "" {
+				//no gateway set, reject this outbound
+				continue
+			}
+
 			outbound = append(outbound, iface.Name)
 			if len(outbound) > 128 {
 				//rules start at 11. 253/254/255 reserved
@@ -466,6 +479,14 @@ func rebuildUplink() {
 
 	if len(outbound) < 2 {
 		//dont need more, outbound will work as is
+
+		if len(outbound) == 1 {
+			//ensure we have the gw set
+			gw, _ := getDefaultGatewayLocked(outbound[0])
+			if gw != "" {
+				exec.Command("ip", "route", "replace", "default", "via", gw, "dev", outbound[0]).Output()
+			}
+		}
 		return
 	}
 
@@ -550,7 +571,7 @@ func rebuildUplink() {
 
 }
 
-func modifyUplinkEntry(ifname string, action string) error {
+func modifyUplinkEntry(ifname, action string, rebuild bool) error {
 	cmd := exec.Command("nft", action, "element", "inet", "filter",
 		"uplink_interfaces", "{", ifname, "}")
 
@@ -581,21 +602,24 @@ func modifyUplinkEntry(ifname string, action string) error {
 		log.Println(cmd)
 	}
 
-	rebuildUplink()
+	if rebuild {
+		rebuildUplink()
+	}
+
 	return err
 }
 
-func addUplinkEntry(ifname string, subtype string) {
+func addUplinkEntry(ifname, subtype string, rebuild bool) {
 	if subtype == "pppup" {
 		//ppp-up connects a ppp to the internt
 		//but packets go directly through the ppp not the pppup provider
 		return
 	}
-	modifyUplinkEntry(ifname, "add")
+	modifyUplinkEntry(ifname, "add", rebuild)
 
 }
-func removeUplinkEntry(ifname string) {
-	modifyUplinkEntry(ifname, "delete")
+func removeUplinkEntry(ifname string, rebuild bool) {
+	modifyUplinkEntry(ifname, "delete", rebuild)
 }
 
 func flushSupernetworkEntries() {
@@ -604,7 +628,7 @@ func flushSupernetworkEntries() {
 	exec.Command("nft", "flush", "set", "ip", "filter", "supernetworks").Output()
 }
 
-func modifySupernetworkEntry(supernet string, action string) {
+func modifySupernetworkEntry(supernet, action string) {
 	cmd := exec.Command("nft", action, "element", "inet", "mangle",
 		"supernetworks", "{", supernet, "}")
 
@@ -1145,8 +1169,59 @@ func applyPrivateNetworkUpstreamDevice(device DeviceEntry) {
 		//if has the tag but not in the verdict map, add it
 		allowPrivateUpstreamAccess(IP)
 	} else if !foundPolicy && inUpstreamAllowed {
-		//if in the verdict map but does not have the tag, remove it
+		//if in the verdict map but does not have the policy, remove it
 		removePrivateUpstreamAccess(IP)
+	}
+}
+
+func hasNoAPIAccess(ip string) bool {
+	cmd := exec.Command("nft", "get", "element", "inet", "filter", "api_block",
+		"{", ip, "}")
+	_, err := cmd.Output()
+	return err == nil
+}
+
+func addNoAPIAccess(ip string) error {
+	cmd := exec.Command("nft", "add", "element", "inet", "filter", "api_block",
+		"{", ip, "}")
+	_, err := cmd.Output()
+
+	if err != nil {
+		log.Println("failed to add element to api_block", err)
+		log.Println(cmd)
+	}
+
+	return err
+}
+
+func removeNoAPIAccess(ip string) error {
+	cmd := exec.Command("nft", "delete", "element", "inet", "filter", "api_block",
+		"{", ip, "}")
+	_, err := cmd.Output()
+
+	if err != nil {
+		log.Println("failed to remove element from api_block", err)
+		log.Println(cmd)
+	}
+
+	return err
+}
+
+func applyNoAPI(device DeviceEntry) {
+	IP := device.RecentIP
+	if IP == "" {
+		return
+	}
+
+	foundPolicy := slices.Contains(device.Policies, DEVICE_POLICY_NOAPI)
+	inNoApi := hasNoAPIAccess(IP)
+
+	if foundPolicy && !inNoApi {
+		//if has the tag but not in the verdict map, add it
+		addNoAPIAccess(IP)
+	} else if !foundPolicy && inNoApi {
+		//if in the verdict map but does not have the policy, remove it
+		removeNoAPIAccess(IP)
 	}
 }
 
@@ -1346,9 +1421,11 @@ func populateSets() {
 			found_wanif = true
 		}
 		if iface.Type == "Uplink" && iface.Enabled == true {
-			addUplinkEntry(iface.Name, iface.Subtype)
+			addUplinkEntry(iface.Name, iface.Subtype, false)
 		}
 	}
+
+	rebuildUplink()
 
 	//As a migration: when no longer in setup mode,
 	//import WANIF into interfaces.json
@@ -2589,15 +2666,26 @@ func getWifiPeers() map[string]string {
 	interfacesConfig := loadInterfacesConfigLocked()
 	Interfacesmtx.Unlock()
 
+	updatePeers := func(iface string) {
+
+		wifi_peers, err := RunHostapdAllStations(iface)
+		if err == nil {
+			for k, peer := range wifi_peers {
+				val, exists := peer["vlan_id"]
+				if exists && (val != "") {
+					peers[k] = iface + "." + peer["vlan_id"]
+				}
+			}
+		}
+
+	}
+
 	for _, entry := range interfacesConfig {
 		if entry.Enabled == true && entry.Type == "AP" {
-			wifi_peers, err := RunHostapdAllStations(entry.Name)
-			if err == nil {
-				for k, peer := range wifi_peers {
-					val, exists := peer["vlan_id"]
-					if exists && (val != "") {
-						peers[k] = entry.Name + "." + peer["vlan_id"]
-					}
+			updatePeers(entry.Name)
+			if len(entry.ExtraBSS) > 0 {
+				for i := range len(entry.ExtraBSS) {
+					updatePeers(entry.Name + ExtraBSSPrefix + strconv.Itoa(i))
 				}
 			}
 		}
@@ -2741,6 +2829,7 @@ func populateVmapEntries(devices map[string]DeviceEntry, groups []GroupEntry, IP
 			log.Println("Unexpected disabled here. Should have aborted earlier")
 		case "lan_upstream":
 			continue //handled in applyPrivateNetworkUpstreamDevice below
+		case "noapi":
 		case "quarantine":
 		case "dns:family":
 		default:
@@ -2750,6 +2839,11 @@ func populateVmapEntries(devices map[string]DeviceEntry, groups []GroupEntry, IP
 
 	//apply other policies
 	applyPrivateNetworkUpstreamDevice(val)
+
+	if !strings.Contains(Iface, ExtraBSSPrefix) {
+		//no api was applied elsewhere.
+		applyNoAPI(val)
+	}
 
 }
 
@@ -2832,6 +2926,11 @@ func establishDevice(devices map[string]DeviceEntry, groups []GroupEntry,
 	// no interface set. abort now
 	if new_iface == "" {
 		return
+	}
+
+	if strings.Contains(new_iface, ExtraBSSPrefix) {
+		//this was a guest wifi network, block API access
+		addNoAPIAccess(entry.RecentIP)
 	}
 
 	exec.Command("ip", "route", "add", routeIP, "dev", new_iface).Run()