+++ installer/scripts/pi-target-install.sh
@@ -113,6 +113,7 @@ ACTION=="add",
 SUBSYSTEM=="usb",
 ENV{ID_VENDOR_ID}=="3574",
 ENV{ID_MODEL_ID}=="6211",
+ENV{DRIVER}!="mt7921u",
 RUN+="/usr/sbin/modprobe mt7921u",
 RUN+="/bin/sh -c 'echo 3574 6211 > /sys/bus/usb/drivers/mt7921u/new_id'"
 EOF
@@ -123,6 +124,7 @@ ACTION=="add",
 SUBSYSTEM=="usb",
 ENV{ID_VENDOR_ID}=="0846",
 ENV{ID_MODEL_ID}=="9060",
+ENV{DRIVER}!="mt7921u",
 RUN+="/usr/sbin/modprobe mt7921u",
 RUN+="/bin/sh -c 'echo 0846 9060 > /sys/bus/usb/drivers/mt7921u/new_id'"
 EOF